Jnuit testing

+++ OnlineFoodDeliveryApp2/src/main/java/com/hexaware/fooddelivery/service/OrdersServiceImp.java
@@ -2,11 +2,17 @@ package com.hexaware.fooddelivery.service;
 
 import java.util.List;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import com.hexaware.fooddelivery.dto.OrdersDTO;
+import com.hexaware.fooddelivery.entity.Customers;
 import com.hexaware.fooddelivery.entity.Orders;
+import com.hexaware.fooddelivery.exception.OrderNotFoundException;
+import com.hexaware.fooddelivery.repository.CustomersRepository;
 import com.hexaware.fooddelivery.repository.OrdersRepository;
 
 @Service
@@ -14,10 +20,21 @@ public class OrdersServiceImp implements IOrdersService {
 	@Autowired
 	OrdersRepository repo;
 	
+	@Autowired
+	CustomersRepository CustomerRepo;
+	
+	
+	Logger logger = LoggerFactory.getLogger(OrdersServiceImpTest.class);
+
 	@Override
 	public Orders addOrders(OrdersDTO ordersDTO) {
-		Orders orders = new Orders();
+		Customers customer=CustomerRepo.findById(ordersDTO.getCustomerId()).orElse(null);
 		
+		
+		
+		
+		Orders orders = new Orders();
+		orders.setCustomer(customer);
 			
 		orders.setCartId(ordersDTO.getCartId());
 		orders.setCustomerId(ordersDTO.getCustomerId());
@@ -25,15 +42,19 @@ public class OrdersServiceImp implements IOrdersService {
 		orders.setDeliveryAddress(ordersDTO.getDeliveryAddress());
 		orders.setPaymentMethod(ordersDTO.getPaymentMethod());
 		orders.setTotalAmount(ordersDTO.getTotalAmount());
-		
-		
+		logger.info("Added to Orders with id"+ordersDTO.getCartId());
+
+
 		return repo.save(orders);
 	}
 
 	@Override
 	public OrdersDTO getById(int cartId) {
 		
-		Orders orders=repo.findById(cartId).orElse(null);
+		Orders orders=repo.findById(cartId).orElse(new Orders());
+		if (orders.getCartId()==0) {
+			  throw new OrderNotFoundException(HttpStatus.NOT_FOUND ,"Orders with cartId:"+cartId+" notfound");
+		}
 		OrdersDTO ordersDTO=new OrdersDTO();
 		
 		ordersDTO.setCartId(orders.getCartId());
@@ -43,12 +64,15 @@ public class OrdersServiceImp implements IOrdersService {
 		ordersDTO.setPaymentMethod(orders.getPaymentMethod());
 		ordersDTO.setTotalAmount(orders.getTotalAmount());
 		
-		
+		logger.info("Get  Orders with id"+orders.getCartId());
+
 		return ordersDTO;
 	}
 
 	@Override
 	public List<Orders> getAllOrders() {
+		logger.info("Fetched all Orders data");
+
 		return repo.findAll();
 	}
 
@@ -63,13 +87,21 @@ public class OrdersServiceImp implements IOrdersService {
 		orders.setDeliveryAddress(ordersDTO.getDeliveryAddress());
 		orders.setPaymentMethod(ordersDTO.getPaymentMethod());
 		orders.setTotalAmount(ordersDTO.getTotalAmount());
+		logger.info("Updated to Orders with id"+ordersDTO.getCartId());
+
 		return repo.save(orders);
 	}
 
 	@Override
 	public void deleteById(int cartId) {
-		Orders orders=repo.findById(cartId).orElse(null);
-		repo.deleteById(orders.getCartId());
+	    Orders orders = repo.findById(cartId).orElse(null);
+
+	    logger.info("Orders deleted");
+
+	    if (orders != null) {
+	        repo.deleteById(orders.getCartId());
+	    }
 	}
+
 
 }
